requirements.txt
# BGN 밝은눈안과 블로그 자동화 시스템 - 필요한 라이브러리

# 웹 프레임워크
streamlit>=1.28.0

# AI 및 자연어 처리
openai>=1.0.0

# 파일 처리
python-docx>=0.8.11
PyPDF2>=3.0.1

# 워드프레스 연동
python-wordpress-xmlrpc>=2.3

# 환경 변수 관리
python-dotenv>=1.0.0

# HTTP 요청
requests>=2.31.0

# 이미지 처리
pillow>=10.0.0

# 데이터 처리
pandas>=2.0.0

# JSON 처리 (내장 모듈이지만 명시)
# json (built-in)

# 날짜 및 시간 처리
# datetime (built-in)

# 파일 경로 처리
# pathlib (built-in)